+++ examples/7.5.3.2-plugin-bufferedupdate-lite-benchmarks-xml.php
@@ -5,6 +5,7 @@ require_once(__DIR__.'/init.php');
 use Solarium\QueryType\Update\Query\Query;
 
 $weight = 'lite';
-$requestFormat = Query::REQUEST_FORMAT_XML;
+$addRequestFormat = Query::REQUEST_FORMAT_XML;
+$delRequestFormat = Query::REQUEST_FORMAT_XML;
 
 require(__DIR__.'/7.5.3-plugin-bufferedupdate-benchmarks.php');