+++ src/Plugin/BufferedAdd/BufferedAddLite.php
@@ -201,7 +201,16 @@ class BufferedAddLite extends AbstractBufferedUpdate
         }
 
         $this->updateQuery->add(null, $command);
-        $this->updateQuery->addCommit($softCommit, $waitSearcher, $expungeDeletes);
+
+        if ($this->updateQuery::REQUEST_FORMAT_CBOR === $this->getRequestFormat()) {
+            $this->updateQuery->addParam('commit', true);
+            $this->updateQuery->addParam('softCommit', $softCommit);
+            $this->updateQuery->addParam('waitSearcher', $waitSearcher);
+            $this->updateQuery->addParam('expungeDeletes', $expungeDeletes);
+        } else {
+            $this->updateQuery->addCommit($softCommit, $waitSearcher, $expungeDeletes);
+        }
+
         $result = $this->client->update($this->updateQuery, $this->getEndpoint());
         $this->clear();