+++ src/Plugin/BufferedDelete/BufferedDeleteLite.php
@@ -9,6 +9,7 @@
 
 namespace Solarium\Plugin\BufferedDelete;
 
+use Solarium\Exception\InvalidArgumentException;
 use Solarium\Exception\RuntimeException;
 use Solarium\Plugin\AbstractBufferedUpdate\AbstractBufferedUpdate;
 use Solarium\Plugin\BufferedDelete\Delete\Id as DeleteById;
@@ -111,6 +112,26 @@ class BufferedDeleteLite extends AbstractBufferedUpdate
     public function getDeletes(): array
     {
         return $this->buffer;
+    }
+
+    /**
+     * Set the request format for the updates.
+     *
+     * Use UpdateQuery::REQUEST_FORMAT_JSON or UpdateQuery::REQUEST_FORMAT_XML as value.
+     *
+     * @param string $requestFormat
+     *
+     * @throws InvalidArgumentException
+     *
+     * @return self Provides fluent interface
+     */
+    public function setRequestFormat(string $requestFormat): self
+    {
+        if ($this->updateQuery::REQUEST_FORMAT_CBOR === strtolower($requestFormat)) {
+            throw new InvalidArgumentException('Unsupported request format: CBOR can only be used to add documents');
+        }
+
+        return parent::setRequestFormat($requestFormat);
     }
 
     /**